Restart DB connection after WebDAV put()
This is similar to what we already do after reassembly.

Two differences:

We check for an ongoing transaction.  If there is one, we bail out
with an error, because reopening the DB connection would break
transaction semantics.

We only check the connection if more than $db_check_interval (default:
300) seconds have passed.  This is to avoid unnecessary DB pings when
uploaded files are small.